+++ test-e2e/cypress-src/plugins/index.ts
@@ -2,5 +2,13 @@ import cypressFailFast = require("../support/cypress-fail-fast/plugin");
 
 export default (on: Cypress.PluginEvents, config: Cypress.PluginConfigOptions): Cypress.ResolvedConfigOptions => {
   cypressFailFast(on, config);
+
+  on("task", {
+    log: function (message) {
+      console.log(message);
+      return null;
+    },
+  });
+
   return config;
 };