+++ CHANGELOG.md
@@ -11,6 +11,19 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
 ### BREAKING CHANGES
 
+## [2.1.0] - 2021-02-21
+
+### Added
+- test(e2e): Check that `test:after:run` event is executed in failed tests.
+
+### Changed
+- feat: Do not stop runner from failed test hook and execute flag task "parallely" in order to let execute test:after:run events. (closes #61)
+- test(e2e): Update Cypress 6 to latest version.
+- chore(deps): Update devDependencies
+
+### Removed
+- test(unit): Remove duplicated test
+
 ## [2.0.0] - 2021-01-17
 
 ### Added